pyproject.toml
[project]
name = "dawn"
version = "0.1.0"
description = "This is a discord bot project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.6.3",
    "google-genai>=1.37.0",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
]
